Add code prevent duplicate data processed at consumer side

+++ src/main/java/com/agh/Payment/utils/CreditCardPayment.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 @Component
 public class CreditCardPayment implements Payment{
     @Override
-    public PaymentResult pay(BigDecimal amount) {
+    public PaymentResult pay(double amount) {
 
         return new PaymentResult(
                 "Credit card payment successful",
